fix:floating-expenses issue

+++ app/modals/ExpenseModel.py
@@ -1,11 +1,11 @@
 from pydantic import BaseModel
 from typing import List
+
+
 class Expense(BaseModel):
-    expenseId:str
-    username:str
-    amount:int
-    category:str
-    description:str
-    expenseDate:str
-
-
+    expenseId: str
+    username: str
+    amount: float
+    category: str
+    description: str
+    expenseDate: str
