+++ app/modals/IncomeModel.py
@@ -1,10 +1,10 @@
 from pydantic import BaseModel
 
+
 class IncomeModel(BaseModel):
-    username:str
-    incomeId:str
-    amount:int
-    category:str
-    description:str
-    incomeDate:str
-    
+    username: str
+    incomeId: str
+    amount: float
+    category: str
+    description: str
+    incomeDate: str
